refactor: reservation.domain이 infrastructure를 import하지 않도록 수정
