+++ src/main/java/roomescape/reservation/infrastructure/JpaReservationViewRepository.java
@@ -3,7 +3,7 @@ package roomescape.reservation.infrastructure;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import roomescape.reservation.domain.ReservationDate;
-import roomescape.reservation.domain.ReservationView;
+import roomescape.reservation.infrastructure.entity.ReservationView;
 
 import java.util.Optional;
