+++ src/main/java/roomescape/reservation/infrastructure/JpaWaitingReservationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.WaitingReservation;
-import roomescape.reservation.infrastructure.vo.MyReservation;
+import roomescape.reservation.domain.vo.MyReservation;
 import roomescape.theme.domain.Theme;
 import roomescape.time.domain.ReservationTime;
